week3/day2/app/requirements.txt
flask>=3.0
gunicorn
Flask-SQLAlchemy
python-dotenv